[Refactor] Move lake late-materialization context to connector
