+++ be/src/connector/lake/lake_global_late_materialization_context.h
@@ -14,51 +14,38 @@
 
 #pragma once
 
-#include <memory>
 #include <optional>
 #include <shared_mutex>
-#include <tuple>
 #include <unordered_map>
 #include <vector>
 
-#include "base/phmap/phmap.h"
 #include "compute_env/query/global_late_materialization_context.h"
 #include "gen_cpp/PlanNodes_types.h"
-#include "storage/rowset/rowset.h"
-#include "storage_primitive/storage_ids.h"
-
-namespace starrocks {
-class Rowset;
-using RowsetSharedPtr = std::shared_ptr<Rowset>;
-} // namespace starrocks
+#include "storage/lake/types_fwd.h"
+#include "storage/rowset/base_rowset.h"
 
 namespace starrocks {
 
-class OlapScanLazyMaterializationContext : public GlobalLateMaterilizationContext {
+class LakeScanLazyMaterializationContext : public GlobalLateMaterilizationContext {
 public:
-    void capture_rowsets(int32_t tablet_id, int64_t version, const std::vector<RowsetSharedPtr>& rowsets);
+    void capture_rowsets(int32_t tablet_id, int64_t version, const std::vector<BaseRowsetSharedPtr>& rowsets);
 
-    RowsetSharedPtr get_rowset(int32_t tablet_id, int32_t dynamic_rssid, int32_t* segment_idx) const;
-
-    using RowsetIdToDRSSId = phmap::parallel_flat_hash_map<RowsetId, uint32_t, HashOfRowsetId>;
-    RowsetIdToDRSSId get_rowset_id_to_drssid(int32_t tablet_id) const;
+    lake::RowsetPtr get_rowset(int32_t tablet_id, int32_t dynamic_rssid, int32_t* segment_idx) const;
+    lake::RowsetPtr get_rowset(const std::vector<lake::RowsetPtr>& rowsets, int32_t drssid, int32_t* segment_idx) const;
 
     int64_t get_rowsets_version(int32_t tablet_id) const {
         std::shared_lock lock(_mutex);
         return _versions.at(tablet_id);
     }
 
-    const TOlapScanNode& scan_node() const { return *_thrift_olap_scan_node; }
-    void set_scan_node(const TOlapScanNode& node);
+    const TLakeScanNode& scan_node() const { return *_thrift_lake_scan_node; }
+    void set_scan_node(const TLakeScanNode& node);
 
 private:
     mutable std::shared_mutex _mutex;
-    using RowsetInfo = std::tuple<RowsetSharedPtr, uint32_t /*dynamic rss id*/>;
-    // tablet_id -> rowset infos
-    std::unordered_map<int32_t, std::vector<RowsetInfo>> _rowsets;
+    std::unordered_map<int32_t, std::vector<lake::RowsetPtr>> _rowsets;
     std::unordered_map<int32_t, int64_t> _versions;
-    std::optional<TOlapScanNode> _thrift_olap_scan_node;
-    size_t _dynamic_rss_id_allocator = 0;
+    std::optional<TLakeScanNode> _thrift_lake_scan_node;
 };
 
 } // namespace starrocks
